mexendo no code

+++ CODE/CONTROL_MOTOR/CONTROL_MOTOR/setDuty_Motor_L.ino
@@ -1,8 +1,8 @@
-void setDuty_Motor_L(float value) {
+void setDuty_Motor_L(double value) {
 
   byte duty;
 
-  value = value / 100;
+  value = value / 100.0;
 
   duty = (value * 256) - 1;
