+++ CODE/CONTROL_MOTOR/CONTROL_MOTOR/setDuty_Motor_R.ino
@@ -1,8 +1,8 @@
-void setDuty_Motor_R(float value) {
+void setDuty_Motor_R(double value) {
 
   byte duty;
 
-  value = value / 100;
+  value = value / 100.0;
 
   duty = (value * 256) - 1;
